# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API ePortfolio",
      "description": "Daftar API ePortfolio",
      "version": "1.0"
    },
    "servers":[
        {
            "url": "http://127.0.0.1:8800/api",
            "description":"localhost"
        },
        {
            "url": "http://103.187.223.15:8800/api",
            "description":"dev"
        }
    ],
...
    "produces": ["application/json"],
    "paths": {
...
      "/post/":{
        "post":{
          "tags":[
            "Post"
          ],
          "summary":"Create Post",
          "description":"Create a new post",
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "mermaidDiagram":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/posts/{id}":{
        "put":{
          "tags":[
            "Post"
          ],
          "summary":"Update Post",
          "description":"Update post",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"post id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    },
                    "desc":{
                      "type":"string"
                    },
                    "mermaidDiagram":{
                      "type":"string"
                    },
                    "isPublic":{
                      "type":"boolean"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Post not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "delete":{
          "tags":[
            "Post"
          ],
          "summary":"Delete Post",
          "description":"Delete post",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"post id"
            }
          ],
          "requestBody":{
            "content":{
              "application/json":{
                "schema":{
                  "type":"object",
                  "properties":{
                    "userId":{
                      "type":"string"
                    }
                  }
                }
              }
            }
          },
          "responses":{
            "200":{
              "description":"Success"
            },
            "403":{
              "description":"Forbidden"
            },
            "404":{
              "description":"Post not found"
            },
            "500":{
              "description": "Error"
            }
          }
        },
        "get":{
          "tags":[
            "Post"
          ],
          "summary":"Get Post",
          "description":"Get post data",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"post id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"Post not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/posts/timeline/all/{id}":{
        "get":{
          "tags":[
            "Post"
          ],
          "summary":"Get All Posts for home",
          "description":"Get all posts data for home, only posts from user who followed",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/posts/all/{id}":{
        "get":{
          "tags":[
            "Post"
          ],
          "summary":"Get All Posts From User",
          "description":"Get all posts data from user",
          "parameters":[
            {
              "in":"path",
              "name": "id",
              "description":"user id"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "404":{
              "description":"User not found"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
      "/posts/search/{keyword}":{
        "get":{
          "tags":[
            "Post"
          ],
          "summary":"Get Post by Keyword",
          "description":"Get post data by keyword",
          "parameters":[
            {
              "in":"path",
              "name": "keyword",
              "description":"keyword"
            }
          ],
          "responses":{
            "200":{
              "description":"Success"
            },
            "500":{
              "description": "Error"
            }
          }
        }
      },
...
    }
 }